Input Splitting
Splits the user's input based on semicolons, thus if the user provides multiple commands in the same input they will all be executed.
Update the input finishing logic to check that the user's last non-white space character is a semicolon, instead of simply checking for the existence of a semicolon.