Left Outer Join Support
The parser now supports parsing SQL's alternative join syntax.
The table now tracks indices of all the tuples in an additional column (which is also cartesian producted). If we are doing a left outer join, once we have selected the appropriate tuples, we find an tuples on the left that haven't been selected and add that tuple with a null right side to the list of selected tuples.